+++ src/main/java/scripting/reactor/ReactorActionManager.java
@@ -177,7 +177,7 @@ public class ReactorActionManager extends AbstractPlayerInteraction {
                 if (d.itemId == 0) {
                     int range = maxMeso - minMeso;
                     int displayDrop = (int) (Math.random() * range) + minMeso;
-                    int mesoDrop = (displayDrop * c.getWorldServer().getMesoRate());
+                    int mesoDrop = (int) (displayDrop * c.getWorldServer().getMesoRate());
                     reactor.getMap().spawnMesoDrop(mesoDrop, reactor.getMap().calcDropPos(dropPos,
                             reactor.getPosition()), reactor, c.getPlayer(), false, (byte) 2, (short) 0);
                 } else {
@@ -193,7 +193,7 @@ public class ReactorActionManager extends AbstractPlayerInteraction {
                 }
             }
         } else {
-            final int worldMesoRate = c.getWorldServer().getMesoRate();
+            final double worldMesoRate = c.getWorldServer().getMesoRate();
 
             dropPos.x -= (12 * items.size());
             short delay = 0;
@@ -201,7 +201,7 @@ public class ReactorActionManager extends AbstractPlayerInteraction {
                 if (d.itemId == 0) {
                     int range = maxMeso - minMeso;
                     int displayDrop = (int) (Math.random() * range) + minMeso;
-                    int mesoDrop = displayDrop * worldMesoRate;
+                    int mesoDrop = (int) (displayDrop * worldMesoRate);
                     MapleMap map = reactor.getMap();
                     map.spawnMesoDrop(mesoDrop, map.calcDropPos(dropPos, reactor.getPosition()), reactor, chr,
                             false, (byte) 2, delay);
