+++ src/main/java/server/loot/FoodDropper.java
@@ -21,13 +21,15 @@ public final class FoodDropper {
     // -- CONFIGURATION --
     // Base chance out of 1,000,000 (1,000,000 = 100%)
     // 5000 = 0.5%
-    private static final int CHANCE_PRIMARY = 2000;       // The main tier for the mob's level
-    private static final int CHANCE_SECONDARY = 1000;     // +/- 1 Tier
-    private static final int CHANCE_TERTIARY = 500;       // +/- 2 Tiers
+    private static final int CHANCE_PRIMARY = 2000; // The main tier for the mob's level
+    private static final int CHANCE_SECONDARY = 1000; // +/- 1 Tier
+    private static final int CHANCE_TERTIARY = 500; // +/- 2 Tiers
 
-    private static final int BOSS_MULTIPLIER = 20;        // 20x chance for bosses
+    private static final int BOSS_MULTIPLIER = 20; // 20x chance for bosses
 
-    public enum Tier { T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, LEGENDARY }
+    public enum Tier {
+        T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, LEGENDARY
+    }
 
     /** Tier -> item ids (quantities default to 1). */
     private static final Map<Tier, int[]> TIER_ITEMS = new EnumMap<>(Tier.class);
@@ -48,9 +50,9 @@ public final class FoodDropper {
     /** Level bands -> tier mapping; tweak as needed. */
     private static final NavigableMap<Integer, Tier> LEVEL_TO_TIER = new TreeMap<>();
     static {
-        LEVEL_TO_TIER.put(30,  Tier.T1);
-        LEVEL_TO_TIER.put(50,  Tier.T2);
-        LEVEL_TO_TIER.put(70,  Tier.T3);
+        LEVEL_TO_TIER.put(30, Tier.T1);
+        LEVEL_TO_TIER.put(50, Tier.T2);
+        LEVEL_TO_TIER.put(70, Tier.T3);
         LEVEL_TO_TIER.put(120, Tier.T4);
         LEVEL_TO_TIER.put(150, Tier.T5);
         LEVEL_TO_TIER.put(200, Tier.T6);
@@ -65,7 +67,9 @@ public final class FoodDropper {
         return (e != null) ? e.getValue() : Tier.T1;
     }
 
-    /** * Rolls for drops independently across multiple tiers centered on the mob's level.
+    /**
+     * * Rolls for drops independently across multiple tiers centered on the mob's
+     * level.
      * EXCLUDES Legendary tier from automatic dropping.
      */
     public static void dropForMonster(
@@ -75,9 +79,9 @@ public final class FoodDropper {
             Point seedPos,
             int rolls,
             byte dropType,
-            boolean playerDrop
-    ) {
-        if (map == null || mob == null || owner == null) return;
+            boolean playerDrop) {
+        if (map == null || mob == null || owner == null)
+            return;
 
         // 1. Identify the "Center" Tier based on level
         Tier primaryTier = tierForLevel(mob.getLevel());
@@ -87,7 +91,7 @@ public final class FoodDropper {
         // 2. Determine Multipliers
         boolean isBoss = mob.getStats().isBoss();
         long bossMulti = isBoss ? BOSS_MULTIPLIER : 1;
-        int serverRate = YamlConfig.config.worlds.get(0).drop_rate;
+        int serverRate = (int) YamlConfig.config.worlds.get(0).drop_rate;
 
         // 3. Iterate through offsets: -2 to +2
         for (int offset = -2; offset <= 2; offset++) {
@@ -108,7 +112,8 @@ public final class FoodDropper {
             // ----------------------------------------
 
             int[] items = TIER_ITEMS.get(currentTier);
-            if (items == null || items.length == 0) continue;
+            if (items == null || items.length == 0)
+                continue;
 
             // Determine Base Chance
             int baseChance;
@@ -129,7 +134,7 @@ public final class FoodDropper {
             for (int i = 0; i < attempts; i++) {
                 if (RNG.nextInt(1_000_000) < finalThreshold) {
                     int itemId = items[RNG.nextInt(items.length)];
-                    Item item = new Item(itemId, (short)0, (short)1);
+                    Item item = new Item(itemId, (short) 0, (short) 1);
                     map.spawnItemDrop((MapObject) mob, owner, item, seedPos, dropType, playerDrop);
                 }
             }
@@ -142,6 +147,11 @@ public final class FoodDropper {
         dropForMonster(map, mob, owner, mob.getPosition(), 1, dropType, false);
     }
 
-    public void setTierItems(Tier tier, int[] itemIds) { TIER_ITEMS.put(tier, itemIds); }
-    public void setLevelBand(int maxLevelInclusive, Tier tier) { LEVEL_TO_TIER.put(maxLevelInclusive, tier); }
+    public void setTierItems(Tier tier, int[] itemIds) {
+        TIER_ITEMS.put(tier, itemIds);
+    }
+
+    public void setLevelBand(int maxLevelInclusive, Tier tier) {
+        LEVEL_TO_TIER.put(maxLevelInclusive, tier);
+    }
 }
